extract formatDate to datehelpers module
