+++ frontend/fingertips-frontend/components/forms/IndicatorSelectionForm/indicatorSelectionForm.test.tsx
@@ -1,9 +1,9 @@
 import { render, screen } from '@testing-library/react';
 import { IndicatorSelectionForm } from '.';
 import { IndicatorDocument } from '@/lib/search/searchTypes';
-import { formatDate } from '@/components/molecules/result';
 import { SearchParams, SearchStateParams } from '@/lib/searchStateManager';
 import { UserEvent, userEvent } from '@testing-library/user-event';
+import { formatDate } from '@/lib/dateHelpers/dateHelpers';
 
 const mockPath = 'some-mock-path';
 const mockReplace = jest.fn();
