--- frontend/fingertips-frontend/lib/dateHelpers/dateComparison.ts
@@ -1,9 +1,0 @@
-export function isWithinOneMonth(
-  currentDate: Date,
-  indicatorDate: Date
-): boolean {
-  const oneMonthAgo = new Date(currentDate);
-  oneMonthAgo.setMonth(oneMonthAgo.getMonth() - 1);
-
-  return oneMonthAgo <= indicatorDate;
-}
