+++ frontend/fingertips-frontend/lib/dateHelpers/dateHelpers.ts
@@ -1,0 +1,18 @@
+export function isWithinOneMonth(
+  currentDate: Date,
+  indicatorDate: Date
+): boolean {
+  const oneMonthAgo = new Date(currentDate);
+  oneMonthAgo.setMonth(oneMonthAgo.getMonth() - 1);
+
+  return oneMonthAgo <= indicatorDate;
+}
+
+export function formatDate(date: Date | undefined): string {
+  if (!date) return 'unknown';
+  const day = String(date.getDate()).padStart(2, '0');
+  const month = date.toLocaleString('en-GB', { month: 'long' });
+  const year = date.getFullYear();
+
+  return `${day} ${month} ${year}`;
+}
